Refactor: 계좌 생성, 그룹 생성 페이지 컨테이너로 구조화하기
