+++ src/app/(pages)/(plain)/account-create/page.tsx
@@ -1,0 +1,1 @@
+export { default } from "@/containers/create/AccountCreateFlow"
